requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
coverage==7.6.1
cryptography==43.0.1
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyQRCode==1.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-jose==3.3.0
pyzbar==0.1.9
qr2text==1.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
